Change command interface
Up to you whether to accept it Coelho. I changed 'halt' to 'stop' in order to prevent accidental shutdown of some Linux systems, as was complained about on some of the threads. It also adds a command not recognized response when it isn't 'stop'.
Note: I only did this on one other repository, so the others would need to be changed.